requirements.txt
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
seaborn>=0.12
pandas>=2.0
pyyaml>=6.0
pytest>=7.0
